Overhauled to make it much faster
Overhauled to make use of "partial_update()".
Also implemented "AutoUpdate" as part of the instantiation of the class, defaults to True, so any changes are written directly.
Calling PaPiRusTextPos(False) will draw nothing automatically unless ".WriteAll()" is called (for more control over when updates are done).

Removing text will now remove it from the image by overwriting the text with a white rectangle.  May need a little work on the positioning/size of the rectangle.